+++ features/sessions/ui/src/commonMain/kotlin/app/campfire/sessions/ui/playback/expanded/ExpandedPlaybackBar.kt
@@ -84,11 +84,13 @@ import app.campfire.core.extensions.fluentIf
 import app.campfire.core.model.Session
 import app.campfire.libraries.api.LibraryItemValidation
 import app.campfire.libraries.api.screen.LibraryItemScreen
+import app.campfire.sessions.ui.composables.OutputDeviceControl
 import app.campfire.sessions.ui.composables.PlaybackSpeedAction
 import app.campfire.sessions.ui.composables.RunningTimerAction
 import app.campfire.sessions.ui.composables.VolumeControl
 import app.campfire.sessions.ui.playback.DefaultNonThemedContentColor
 import app.campfire.sessions.ui.playback.DefaultNonThemedSheetColor
+import app.campfire.sessions.ui.playback.OutputDeviceUiState
 import app.campfire.sessions.ui.playback.PlaybackUiState
 import app.campfire.sessions.ui.playback.PlayerUiEvent
 import app.campfire.sessions.ui.playback.PlayerUiState
@@ -175,6 +177,7 @@ internal fun <T> T.ExpandedPlaybackBar(
       syncState = playbackState.syncUiState,
       playbackHistoryEnabled = playbackState.playbackHistoryEnabled,
       volumeState = playbackState.volume,
+      outputDeviceState = playbackState.outputDevices,
       onClose = onClose,
       sharedTransitionScope = this,
       animatedVisibilityScope = this,
@@ -195,6 +198,7 @@ internal fun ExpandedPlaybackBar(
   syncState: SyncUiState,
   playbackHistoryEnabled: Boolean,
   volumeState: VolumeUiState?,
+  outputDeviceState: OutputDeviceUiState?,
 
   onClose: () -> Unit,
   sharedTransitionScope: SharedTransitionScope,
@@ -360,6 +364,7 @@ internal fun ExpandedPlaybackBar(
             itemValidation = itemValidation,
             playbackHistoryEnabled = playbackHistoryEnabled,
             volumeState = volumeState,
+            outputDeviceState = outputDeviceState,
             onClose = onClose,
             windowSizeClass = windowSizeClass,
             animatedVisibilityScope = animatedVisibilityScope,
@@ -384,6 +389,7 @@ private fun SharedTransitionScope.ExpandedPlaybackContent(
   itemValidation: LibraryItemValidation,
   playbackHistoryEnabled: Boolean,
   volumeState: VolumeUiState?,
+  outputDeviceState: OutputDeviceUiState?,
 
   onClose: () -> Unit,
 
@@ -622,6 +628,9 @@ private fun SharedTransitionScope.ExpandedPlaybackContent(
       },
       volumeContent = {
         volumeState?.let { VolumeControl(state = it) }
+      },
+      outputDeviceContent = {
+        outputDeviceState?.let { OutputDeviceControl(state = it) }
       },
       onEqualizerClick = {
         scope.launch {